+++ src/components/Menu/index.js
@@ -1,0 +1,3 @@
+import NavBar from './NavBar';
+
+export { NavBar };
