+++ src/components/common/index.js
@@ -3,7 +3,6 @@ import FormInput from './FormInput';
 import List from './List';
 import LoadingComponent from './LoadingComponent';
 import Button from './Button';
-import NavBar from './Menu';
 // notice we're building out a 'package' of reusables here and exporting them as an object of component properties.
 // to use this, simply `import {foo, bar, baz} from '<path-to-this-directory>/ReusableComponents';`
-export { FormButton, FormInput, List, LoadingComponent, Button, NavBar };
+export { FormButton, FormInput, List, LoadingComponent, Button };
